+++ src/StepperISR.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 
 #include "FastAccelStepper.h"
-#include "common.h"
+#include "fas_common.h"
 
 // Here are the global variables to interface with the interrupts